modified

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: kamako <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/20 17:57:56 by kamako            #+#    #+#             */
-/*   Updated: 2019/05/21 16:31:45 by kamako           ###   ########.fr       */
+/*   Updated: 2019/05/27 10:30:37 by kamako           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *s)
+size_t		ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
